reuse text view component
